+++ src/main/java/chess/domain/board/Board.java
@@ -1,8 +1,9 @@
 package chess.domain.board;
 
-import chess.domain.*;
 import chess.domain.piece.Direction;
 import chess.domain.piece.Piece;
+import chess.domain.piece.Role;
+import chess.domain.square.*;
 
 import java.util.ArrayList;
 import java.util.List;
