+++ src/main/java/chess/domain/piece/BlankPiece.java
@@ -1,7 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.Role;
-import chess.domain.Side;
+import chess.domain.square.Side;
 
 public class BlankPiece extends Piece {
     public BlankPiece(final Side side, final Role role) {
