+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,8 +1,7 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
-import chess.domain.Role;
-import chess.domain.Side;
+import chess.domain.square.Color;
+import chess.domain.square.Side;
 
 public abstract class Piece {
     protected final Side side;
