+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,7 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.Role;
-import chess.domain.Side;
+import chess.domain.square.Side;
 
 public class Rook extends Piece {
     public Rook(final Side side, final Role role) {
