+++ src/test/java/chess/domain/piece/PawnTest.java
@@ -1,8 +1,7 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
-import chess.domain.Role;
-import chess.domain.Side;
+import chess.domain.square.Color;
+import chess.domain.square.Side;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -11,9 +10,9 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
 
-import static chess.domain.Color.BLACK;
-import static chess.domain.Color.WHITE;
 import static chess.domain.piece.Direction.*;
+import static chess.domain.square.Color.BLACK;
+import static chess.domain.square.Color.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class PawnTest {
